- Moved clAuiBook into its own file - Added supprt for PageClosing, PageChanging and Context menu events
