+++ Plugin/Notebook.h
@@ -28,8 +28,6 @@
 
 #include "cl_defs.h"
 
-#include <wx/aui/auibook.h>
-
 #if CL_USE_NATIVEBOOK
 
 #include "clNativeNotebook.hpp"
@@ -40,25 +38,5 @@ typedef clNativeNotebook Notebook;
 typedef clGenericNotebook Notebook;
 
 #endif // CL_USE_NATIVEBOOK
-
-class WXDLLIMPEXP_SDK clAuiBook : public wxAuiNotebook
-{
-public:
-    clAuiBook(wxWindow* parent, wxWindowID id, const wxPoint& pos = wxDefaultPosition,
-              const wxSize& size = wxDefaultSize, long style = 0);
-    ~clAuiBook() override;
-    size_t GetAllTabs(clTabInfo::Vec_t& tabs);
-    clTabHistory::Ptr_t GetHistory() const { return m_history; }
-    void MoveActivePage(int newIndex);
-    int SetSelection(size_t newPage) override;
-    int ChangeSelection(size_t n) override;
-
-protected:
-    void OnPageClosed(wxAuiNotebookEvent& event);
-    void OnPageChanged(wxAuiNotebookEvent& event);
-
-private:
-    clTabHistory::Ptr_t m_history;
-};
 
 #endif // NOTEBOOK_H
